Modify the framework so that it works after refactoring
1) casefileparser.py: Import module "string", which was missed
2) mapper.py, proxy.py: We plans to only support Python, so
the codes about "ruby" and "Java" should be removed, but didn't
do it in this patch, as the mainly purpose of this series patches
are to reorgnize the source code trees. What the patch does is
to hardcode "language" to "Python" as a tempory solution. Will
refine "mapper.py" and "proxy.py" once this patch series are pushed.