repo: Add test for console input and output operations
This test checks if the console input and output work as desired. The
test takes as an argument a filename, whose contents are sent to the
console. The optional parameter 'expect' can be set to a filename that
should contain expected output from the guest and is compared with the
actual output. The optional parameter 'output' can be set to a filename
where the test script saves the output of the host (useful for initial
test setup).

This test requires that the guest machine runs code that handles input and
output on the serial console (programs such as agetty or something like that).