test-API: Destroy the redundant usage comments and function
Since we will generate the usage string automatically with
the newly introduced util function usage(reqs, opts), the
redundant comments and usage functions in all the scripts
under repos/ can be destroyed now.

And the comments on the top of each scripts to explain
what the scripts are improved/shortened, the format is
changed to use "#" instead of "'''".