test-API: Add required_params and optional_params to repo scripts
While adding optional_params, I found we need the dict for it, but
not a tuple, as optional params could have default values. So there
will be a follow up patch to change utils.usage(..., ...). And we
need to figure out the default values for the optional params in
each script, also get rid of the hard-codes to set the default
values. But all of these will be in seperate patches in future.