Remove the useless scripts directory and script under it
The script is useful only if all the scripts have good document
comments. Unfortunately, it's not true, and it's burden for
developers to write those documents in specific format somehow,
on the other hand, because we introduced required_opts, optional_opts
for each cases to generate the usage strings, and new option "-t"
to print template for cases, it's no need to write documents
in specific format for scripts now.