Rename and modify the syntax of BUGSKIP
Skipping only step(s) of a test case doesn't make sense, e.g.
there is a bug of domain defining, if we only skip the "domain:define"
in all test cases, that means all the followup steps (e.g. domain
start, checking) of all of the test cases use "domain:define" will
run forward blindly. That's really not correct.

This patch modify the syntax so that it skip the whole test case
due to either a bug or lacking of testing environment. And thus the
name is changed into "SKIP".

No codes to support the new syntax yet, but it's fine, as we still
don't support to execute all test cases in a bundle. And the syntax
to skip step(s) in test cases should be just destroyed. There will
be follow up patches to destroy the codes to skip test steps in
test cases, and patches to support to execute all test cases
in a bundle and the new skipping syntax.