Add getCellsFreeMemory test case
It is hard to check if the free memory for each node is right, because
memory change too fast in most machine (even machine in low load).
Get this free memory 1000 times and get the Standard Deviation via
get_standard_deviation().

177 is a expectations get from 1000 times test in a low load machine,
so this case may still get fail in some big machine.

Signed-off-by: Luyao Huang <lhuang@redhat.com>